[Refactor] 패키지 네이밍 변경 #16 - common -> global - config -> global/config
