+++ src/main/java/com/wooringpang/userservice/global/dto/CommonSearchCondition.java
@@ -1,4 +1,4 @@
-package com.wooringpang.userservice.common.dto;
+package com.wooringpang.userservice.global.dto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
